wifi_nxp: fix aarch64 compile warnings with hostap enabled
Fix additional aarch64 related compile warnings in wifi_nxp
when hostap is enabled. Replace integer casts with uintptr_t
and fix sizeof() format specifiers.

- Replace (void *)bss_type with
  (void *)(uintptr_t)bss_type in wifi_event_completion calls
- Replace (t_u32)(msg->data) with (uintptr_t)(msg->data)
- Replace %d with %zu for sizeof() format strings
- Fix cur_network_idx cast via uintptr_t

+++ mcux/middleware/wifi_nxp/wlcmgr/wlan.c
@@ -6616,7 +6616,7 @@ static void wlcm_request_connect(struct wifi_message *msg, enum cm_sta_state *ne
     ret = do_connect((int)(uintptr_t)msg->data);
 #else
     wlan.scan_count      = 0;
-    wlan.cur_network_idx = (int)msg->data;
+    wlan.cur_network_idx = (int)(uintptr_t)msg->data;
 #if !CONFIG_WIFI_NM_WPA_SUPPLICANT
     if (wlan.roam_reassoc == true)
     {
